data_entry script updated: For college node, creating private group after it's creation.
